python_bot/requirements.txt
fastapi
uvicorn
requests
beautifulsoup4
google-genai
supabase
apscheduler
pycryptodome
python-dotenv
pydantic
httpx
python-multipart
cloudscraper>=1.2.71
curl_cffi>=0.14.0